requirements.txt
open-interpreter==0.3.4